requirements.txt
pyautogui
pillow
pynput
opencv-python
numpy
pillow
mss
pyfiglet
colorama
discord-webhook